requirements.txt
alembic==1.12.1
blinker==1.7.0
click==8.1.7
flask==3.0.0
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
greenlet==3.0.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
typing-extensions==4.8.0
werkzeug==3.0.1
zipp==3.17.0
